Small fix to schema

+++ src/main/db/create-schema.sql
@@ -139,8 +139,8 @@ create table unique_titles (
   id int not null auto_increment,
   name varchar(80) not null,
   unique_distinction_id int not null,
-  treshold double,
-  twinoid_points int,
+  treshold int,
+  twinoid_points double,
   primary key (id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8;
 alter table unique_titles add constraint fk_unique_titles_unique_distinctions_unique_distinction_id foreign key (unique_distinction_id) references unique_distinctions(id);
